Modified utility classes to support 32/64 bit building of libtriton
git-svn-id: file:///Users/charlie/neptune/trunk@163 9f0c2d39-262c-0410-996f-cacf60396088

+++ kernel/mem/allocator.d
@@ -23,7 +23,7 @@ class FixedAllocator
         free = null;
     }
 
-    new(ulong size, void* pos)
+    new(size_t size, void* pos)
     {
         return pos;
     }
